Cleaned up warnings, upgraded project to Xcode 4.6.2 and Aperture SDK 3.4. Only building for 10.8 right now.

+++ ApertureToGallery3/JSON/SBJsonStreamParser.m
@@ -123,7 +123,7 @@
 
 - (void)handleObjectStart {
 	if (depth >= maxDepth) {
-		self.error = [NSString stringWithFormat:@"Parser exceeded max depth of %lu", maxDepth];
+		self.error = [NSString stringWithFormat:@"Parser exceeded max depth of %lu", (unsigned long)maxDepth];
 		states[depth] = kSBJsonStreamParserStateError;
 		
 	} else {
@@ -136,7 +136,7 @@
 }
 - (void)handleArrayStart {
 	if (depth >= maxDepth) {
-		self.error = [NSString stringWithFormat:@"Parser exceeded max depth of %lu", maxDepth];
+		self.error = [NSString stringWithFormat:@"Parser exceeded max depth of %lu", (unsigned long)maxDepth];
 		states[depth] = kSBJsonStreamParserStateError;
 	} else {
